requirements.txt
advocate==1.0.0
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
celery-redbeat==2.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.10
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-cors-headers==3.10.1
django-filter==21.1
django-mjml==0.11.0
django-timezone-field==4.2.1
djangorestframework==3.13.1
drf-jwt==1.19.1
drf-spectacular==0.21.1
Faker==11.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.0.1
jsonschema==4.3.2
kombu==5.2.3
Markdown==3.3.6
msgpack==1.0.3
ndg-httpsclient==0.5.1
netifaces==0.11.0
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
regex==2021.11.10
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
mysqlclient==2.1.0
# psycopg2
djangorestframework-simplejwt
django-redis
django-celery-results==2.2.0
sqlalchemy